Splitted LaMetric stuff in separate repo for late reuse, rearranged classes
