--- src/main/java/de/ketrwu/lametric/JacksonUtils.java
@@ -1,17 +1,0 @@
-package de.ketrwu.lametric;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class JacksonUtils {
-
-    private static ObjectMapper objectMapper = new ObjectMapper();
-
-    static {
-        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-    }
-
-    public static ObjectMapper getObjectMapper() {
-        return objectMapper;
-    }
-}
