--- src/main/java/de/ketrwu/lametric/entity/lametric/LaMetricRequest.java
@@ -1,17 +1,0 @@
-package de.ketrwu.lametric.entity.lametric;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class LaMetricRequest {
-
-    private String deviceId;
-    private String appInstance;
-    private String awsRequestId;
-    @JsonIgnore
-    private String authorization;
-
-}
