+++ src/main/java/de/ketrwu/lametric/service/SpotifyService.java
@@ -3,17 +3,17 @@ package de.ketrwu.lametric.service;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
-import de.ketrwu.lametric.JacksonUtils;
-import de.ketrwu.lametric.Logger;
-import de.ketrwu.lametric.entity.lametric.LaMetricRequest;
-import de.ketrwu.lametric.entity.spotify.SpotifyCurrentlyPlaying;
-import de.ketrwu.lametric.entity.spotify.SpotifyEndpoint;
+import de.ketrwu.lametric.entity.SpotifyCurrentlyPlaying;
+import de.ketrwu.lametric.entity.SpotifyEndpoint;
+import de.ketrwu.lametric.lambda.LaMetricJacksonUtils;
+import de.ketrwu.lametric.lambda.LaMetricJsonLogger;
+import de.ketrwu.lametric.lambda.entity.lametric.LaMetricRequest;
 
 import java.io.IOException;
 
 public class SpotifyService {
 
-    private static final Logger LOG = new Logger(SpotifyService.class);
+    private static final LaMetricJsonLogger LOG = new LaMetricJsonLogger(SpotifyService.class);
     private static final String API_URL_BASE = "https://api.spotify.com/v1";
 
     private static <T> T sendRequest(LaMetricRequest request, Class<? extends T> clazz) throws UnirestException, IOException {
@@ -33,7 +33,7 @@ public class SpotifyService {
                 return null;
             }
 
-            T body = JacksonUtils.getObjectMapper().readValue(response.getBody(), clazz);
+            T body = LaMetricJacksonUtils.getObjectMapper().readValue(response.getBody(), clazz);
 
             if (response.getStatus() >= 200 && response.getStatus() < 300) {
                 LOG.info(request, "Spotify ok", body);
